requirements.txt
aws-cdk-lib==2.42.0
constructs>=10.0.0,<11.0.0

boto3~=1.26.15
awswrangler~=2.17.0
botocore~=1.29.15
pandas~=1.5.2
requests~=2.28.1